% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aliases.R
\name{aliases}
\alias{aliases}
\title{This file contains aliases for the more complicated names of our
other functions.}
\description{
This file contains aliases for the more complicated names of our
other functions.
}

