% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aliases.R, R/buildConcordance.R
\name{concordance}
\alias{buildConcordance}
\alias{concordance}
\title{Build a concordance for a keyword}
\usage{
concordance(dt, keyword, context)

buildConcordance(dt, keyword, context)
}
\arguments{
\item{dt}{A \code{docTexts} object that contains your collection
of texts.}

\item{keyword}{The keyword that you are studying.}

\item{context}{The number of words on each side of the
       \code{keyword} to search for.}
}
\value{
A \code{docConcordance} object: essentially a structured list that contains
every use of the keyword in the collection.
}
\description{
Build a concordance for a keyword
}
\section{What it does}{

Builds a concordance for a keyword, using keyword-in-context (KWIC) technique.
The first parameter is a \code{docTexts} object that contains all of your texts.
The second parameter is the keyword you are interested studying. The third gives the
size of the context
that you want to use for the concordance: i.e., how many words on each side of
your keyword that you want to look at. The context is sometimes referred to as the
'search window'.
}
\examples{
dc = buildConcordance(dt, "justice", 5)
}

