% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildDocList.R
\name{buildDocList}
\alias{buildDocList}
\title{}
\usage{
buildDocList(directory, stopwordsFile = "", indexFile = "",
  includeTexts = FALSE, wizard = FALSE)
}
\arguments{
\item{directory}{A string that is the path to the directory where the
files that make up your corpus are located.}

\item{stopwordsFile}{}

\item{indexFile}{}

\item{includeTexts}{}

\item{wizard}{}
}
\description{

}
\examples{
Placeholder
}

