% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{buildModel}
\alias{buildModel}
\title{Build a topic model}
\usage{
buildModel(dl, tnum, train = 100, maximize = 10, listLength = 100)
}
\arguments{
\item{dl}{A \code{docList} object, representing your collection (must include
a stopwords file and must have imported texts).}

\item{tnum}{The number of topics you want \code{mallet} to search for.}

\item{train}{The number of times \code{mallet} will iterate over your
collection.}

\item{maximize}{A few extra iterations meant to pick the best topic for each word.}

\item{listLength}{The number of words from each topic you'd like to see displayed.}
}
\value{
dmod           The completed \code{docModel} object for viewing and simple analysis.
}
\description{
Run \code{mallet} over a \code{tei2r} collection.
}
\section{Description}{

Trains a basic topic model using the 'mallet' package and saves the output
into a single \code{docModel} object. For easy viewing, the resulting object 
contains a data frame of each topic's top words and each document's most
prevalent topics.
}
\examples{
dmod = buildModel(dl = dl, tnum = 50)
View(dmod@topics) # To view a table showing the words of each topic
View(dmod@frequencies) # To view the frequency of the topics in each document
}

