% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{Convert a plain-text or TEI document into a character vector}
\usage{
cleanup(filepath, stopwords = c(), normalize = TRUE)
}
\arguments{
\item{filepath}{A path to the file that will be converted.}

\item{normalize}{A logical condition. If "TRUE", text will be converterd to 
all lower case and stopwords will be removed. Also, all 
instances of '∫' and 'ſ' will be convertedto 's', all 
numeric characters will be removed, 'vv' will be converted to 'w', 
and ''d' and ''ring' will be converted to 'ed' and 'ering'
respectively, and all special characters will be removed.}
}
\description{
Convert a plain-text or TEI document into a character vector
}
\examples{
locke.path = "~/Desktop/locke2ndTreatise.txt"
cleanup(locke.path, removeCaps = TRUE, removeStopwords = FALSE)
}

