% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docConcordance.R
\docType{class}
\name{docConcordance-class}
\alias{docConcordance}
\alias{docConcordance-class}
\title{An S4 class that holds a concordance for one term}
\description{
\code{docConcordance} creates a special \code{tei2r} object called a concordance
}
\section{Slots}{

\describe{
\item{\code{directory}}{A string that gives the filepath to the main directory}

\item{\code{indexFile}}{A string providing the filepath to the index (usually a .csv file)}

\item{\code{keyword}}{A string providing the keyword}

\item{\code{context}}{An integer providing the context window, before and after the keyword}

\item{\code{concordance}}{A list of "keyword-in-context" character vectors.}
}}
\section{What it does}{

The docConcordance object is a structured list that holds all uses of a keyword
found in a collection.
}

