% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docFrequencies.R
\docType{class}
\name{docFrequencies-class}
\alias{docFrequencies}
\alias{docFrequencies-class}
\title{An S4 class that contains and manages word-frequency data}
\description{
\code{docFrequencies} creates a special \code{tei2r} object called
}
\section{Slots}{

\describe{
\item{\code{directory}}{A string ...}

\item{\code{index}}{A string [NOTE: THIS SHOULD BE CALLED indexFile.}

\item{\code{straight}}{A list ... [RENAME 'RAW']}

\item{\code{proportional}}{A list ...}
}}
\section{What it does}{

Here we can paste longer descriptions of what each function does.
}

