% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docList.R
\docType{class}
\name{docList-class}
\alias{docList}
\alias{docList-class}
\title{An S4 class to represent a document collection}
\description{
\code{docList} returns a special \code{tei2r} object called ...
}
\section{Slots}{

\describe{
\item{\code{directory}}{A string that gives the filepath to the main directory (folder)}

\item{\code{filenames}}{A vector ...}

\item{\code{paths}}{A vector...}

\item{\code{indexFile}}{A string...}

\item{\code{index}}{A data frame...}

\item{\code{stopwordsFile}}{A string...}

\item{\code{stopwords}}{A vector...}

\item{\code{text}}{A list (EXCEPT WE'RE NOT USING 'TEXT' SLOT.)}
}}
\section{What it does}{

Here we can paste longer descriptions of what each function does.
}

