% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docMatrix.R
\docType{class}
\name{docMatrix-class}
\alias{docMatrix}
\alias{docMatrix-class}
\title{An S4 class that contains and manages word-frequency data}
\description{
\code{docMatrix} creates a special \code{tei2r} object called
}
\section{Slots}{

\describe{
\item{\code{directory}}{A string ...}

\item{\code{indexFile}}{A string [NOTE: THIS SHOULD BE CALLED indexFile.}

\item{\code{type}}{A string ... [RENAME 'RAW']}

\item{\code{mat}}{A matrix ...}
}}
\section{What it does}{

Here we can paste longer descriptions of what each function does.
}

