% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docModel.R
\docType{class}
\name{docModel-class}
\alias{docModel}
\alias{docModel-class}
\title{An S4 class to represent a topic model.}
\description{
\code{docModel} contains
a structured list of information about a topic model.
}
\section{Slots}{

\describe{
\item{\code{directory}}{A string that gives the filepath to the main directory.}

\item{\code{index}}{A data frame that holds the meta-data for each document in
the collection.}

\item{\code{topics}}{A data frame that holds the most frequent words in each topic, for easy viewing.}

\item{\code{frequencies}}{A matrix showing how frequent each topic is in each document.}

\item{\code{weights}}{A list showing how heavily weighted each word is in each topic.}

\item{\code{termMatrix}}{A complete matrix of every word in the collection, and its representation
                     in each topic (rarely consulted directly).}
}}
\section{What it does}{

The \code{docModel} object provides an easy way of handling R objects
creating while using the \code{mallet} package. It holds an index of your files along with
all of the results from the model, which are formatted for easy viewing.
}
\seealso{
buildModel
}

