% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDocFrequencies.R
\name{frequencies}
\alias{frequencies}
\title{Calculate word frequencies across a collection}
\usage{
frequencies(dt)
}
\arguments{
\item{dt}{A \code{docTexts} object that contains the full
text of each document in your corpus.}

\item{limit}{A numeric value that limits the number of words
               to find the frequency of in each document in
               \code{dt}. Defaults to all terms.}
}
\value{
df     a \code{docFrequencies} object with frequency
               data for the corpus.

Returns a docFrequencies object.
}
\description{
Get the frequencies of all words in a collection.
}
\section{What it does}{

Get term frequency data for the terms in each document across the
collection, represented by a \code{docTexts} object.  Also
determines the frequency and proportional
frequency of all terms across the corpus. The first parameter is a
\code{docTexts} object containing the full cleaned text for each document
in your colleciton.  The second is a limit on the number of terms to
include from each document in the frequency lists, in case what you're
looking for are only the most frequent words.
}
\examples{
df = getDocFrequencies(dt)
df = getDocFrequencies(dt, 500)
df = getDocFrequencies(dt=dt, limit=1000)
}

