% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/writeDocFrequencies.R
\name{frequencies2csv}
\alias{frequencies2csv}
\title{Export word frequencies to .csv}
\usage{
frequencies2csv(df, filename, limit = 100)
}
\arguments{
\item{df}{The \code{docFrequencies} object that holds your frequency values.}

\item{filename}{The name of the file you want the frequency tables to be saved to.
\strong{requires} that the file eextension be a .csv file.  Stored in the \code{directory}
provided in the \code{docList} object.}

\item{limit}{The number of words to include from each document.}
}
\description{
This function will write a \code{docFrequencies} object to a comma
seperated values \code{.csv} file.  The values are derived from a
\code{docFrequencies} object and use the \code{raw} slot.  Each
document will have a seperate set of columns that are the word
and frequency.  Each set of columns is labeled with the document's
ID that is derived from the index file.  The file is stored in the
directory that was provided when you built your \code{docList}.
}

