% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDocFrequencies.R
\name{getDocFrequencies}
\alias{getDocFrequencies}
\title{Placeholder}
\usage{
getDocFrequencies(dt, limit = 0)
}
\arguments{
\item{dt}{}

\item{limit}{}
}
\description{
Placeholder
}
\examples{
placeholder
}

