% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aliases.R, R/getDocTexts.R
\name{texts}
\alias{getDocTexts}
\alias{texts}
\title{Get the texts of a collection}
\usage{
texts(dl)

getDocTexts(dl, removeCaps = TRUE, removeStopwords = TRUE,
  normalizeLongS = TRUE)
}
\arguments{
\item{dl}{The \code{docList} object that contains the \code{index} with
          the \code{paths} to the files for each text.}

\item{removeCaps}{Option to remove capitalization. Default is 'TRUE'.}

\item{removeStopwords}{Option to remove stopwords, stored in the corresponding \code{docList} object. Default is 'TRUE'.}

\item{normalizeLongS}{Option to normalize '∫' and 'ſ' to 's'. Default is 'TRUE'.}
}
\value{
dt The \code{docTexts} object that contains the texts of the corpus,
           the path to the \code{indexFile} and the original \code{directory}
           that the \code{docList} object was built from.
}
\description{
Imports a collection of documents into R and performs basic text processing
}
\section{What it does}{

This function collects, cleans up, and stores
the text of the collection's documents in a single object. The texts are held as
vectors in a single list, labeled by the corresponding id
in the index file.  The id is usually the filename or the Text Creation Partnership number.
}
\examples{
dt = texts(dl)
dt = getDocTexts(dl)
}
\seealso{
docTexts
}

