% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFiles.R
\name{getFiles}
\alias{getFiles}
\title{This function acts as a wizard for the user for downloading
files from the EEBO TCP based on a number of available filters.}
\usage{
getFiles(src = "TCP", ids = "", directory = "", dl = NA, append = F)
}
\arguments{
\item{src}{A string that determines the source from which to
              download files.  Defaults to TCP and currently only
              functions for TCP.}

\item{ids}{One or more TCP id numbers that will be downloaded.}

\item{dl}{A \code{docList} object that will be modified after
              files are downloaded.}

\item{append}{A boolean value that determines whether or not the
              function should append values to the \code{index}
              of \code{dl} or not. (Usually used to determine
              if the function is being called by another function)}
}
\value{
dl   A modified version of \code{dl}.  Most modifications
              are made to the index.  If not in append mode,
              the \code{directory} will be changed to
              \code{directory}/data/ and the \code{indexFile} will
              also be in this location.
}
\description{
This function acts as a wizard for the user for downloading
files from the EEBO TCP based on a number of available filters.
}
\examples{
dl = getFiles()
dl = getFiles(ids=c("idOne", "idTwo"))
dl = getFiles(dl=dl)
dl = getFiles(dl=dl, append=T)
}

