% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTexts.R
\name{importTexts}
\alias{importTexts}
\title{Import a collection of texts into R}
\usage{
importTexts(dl, normalize = TRUE)
}
\arguments{
\item{dl}{The \code{docList} object that contains the \code{index} with
the \code{paths} to the files for each text.}

\item{normalize}{A logical condition. If "TRUE", text will be converterd to 
all lower case and stopwords will be removed. Also, all 
instances of '∫' and 'ſ' will be convertedto 's', all 
numeric characters will be removed, 'vv' will be converted to 'w', 
and ''d' and ''ring' will be converted to 'ed' and 'ering'
respectively, and all special characters will be removed.}
}
\value{
dl The \code{docList} object that contains the texts of the corpus,
           the path to the \code{indexFile} and the original \code{directory}
           that the \code{docList} object was built from.
}
\description{
Imports a collection of documents into R and performs basic text processing
}
\section{What it does}{

This function collects, cleans up, and stores
the text of the collection's documents in a single object. Essentially, it
runs the \code{\link{cleanup}} function over a folder of documents. The texts are held as
vectors in a single list, labeled by the corresponding id 
in the index file.  The id is usually the filename or the Text Creation Partnership number.
}
\examples{
dt = importTexts(dl)
}
\seealso{
cleanup
}

