% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natlaw.R
\docType{data}
\name{natlaw}
\alias{natlaw}
\title{Data - Collection: 17th Cent. Political Theory}
\format{A docList with an index of 33 items.}
\usage{
natlaw
}
\description{
A \code{docList} object containing metadata for a sample dataset
of 33 17th-century documents relating loosely to political theory.
}
\section{What it contains}{

\itemize{
\item A26092  Astell, Mary. 1694. A serious proposal to the ladies, for the advancement of their true and greatest interest by a lover of her sex.
\item A28298  Blackmore, Richard. 1695. Prince Arthur an heroick poem in ten books
\item A29968	Buchanan, George. 1642. Tyrannicall-government anatomized, or, A discovrse concerning evil-councellors.
\item A30402	Burnet, Gilbert. 1700. Reflections on a book entituled (The rights, powers, and privileges of an English convocation)
\item A30483	Burnet, Thomas. 1697. Remarks upon An essay concerning humane understanding.
\item A53044	Newcastle, Margaret Cavendish. 1668. The description of a new world, called the blazing-world.
\item A38258	Charles I, King of England. 1648. Eikōn basilikē, The pourtraicture of His Sacred Majestie in his solitudes and sufferings.
\item A36581	Dryden, John. 1681. Absalom and Achitophel a poem.
\item A36643	Dryden, John. 1692. MacFlecknoe
\item A41307	Filmer, Robert. 1696. Observations concerning the original and various forms of government as described.
\item A41308	Filmer, Robert. 1680. Patriarcha, or, The natural power of Kings.
\item A42231	Grotius, Hugo. 1675. Of the government and rites of the ancient church.
\item A42234	Grotius, Hugo. 1655. Of the law of warre and peace.
\item A85746	Grotius, Hugo. 1651. Of the authority of the highest powers about sacred things. Or, The right of the state in the Church.
\item A87137	Harrington, James. 1657. The prerogative of popular government. A politicall discourse in two books.
\item A45613	Harrington, James. 1656. The common-wealth of Oceana.
\item A43978	Hobbes, Thomas. 1652. De corpore politico, or, The elements of law, moral and politick.
\item A43998	Hobbes, Thomas. 1651. Leviathan.
\item A48901	Locke, John. 1690. Two treatises of government.
\item A48884	Locke, John. 1689. A letter concerning toleration.
\item A50274	Machiavelli, Niccolò. 1680. The works of the famous Nicholas Machiavel.
\item A50883	Milton, John. 1644. Areopagitica.
\item A50893	Milton, John. 1692. A defence of the people of England.
\item A45612	Harrington, James. 1660. The Censure of the Rota upon Mr Miltons book, entituled, The ready and easie way to establish a free common-wealth.
\item A50898	Milton, John. 1650. Eikonoklestēs.
\item A50919	Milton, John. 1667. Paradise lost.
\item A50948	Milton, John. 1660. The readie and easie vvay to establish a free commonwealth.
\item A50949	Milton, John. 1641. The reason of church-government urg'd against prelaty.
\item A56253	Pufendorf, Samuel. 1695. An introduction to the history of the principal kingdoms and states of Europe
\item A59474	Shaftesbury, Anthony Ashley Cooper. 1675. A letter from a Parliament man to his friend.
\item A59475	Shaftesbury, Anthony Ashley Cooper. 1675. A letter from a person of quality to his friend in the country.
\item A30442	F. B. 1698. A free but modest censure on the late controversial writings and debates.
\item A61556	Stillingfleet, Edward. 1680. The grand question, concerning the Bishops right to vote in Parliament.
}
}
\examples{
# Load the collection, then import the texts
data(natlaw)
dt = importTexts(natlaw)


}
\references{
https://github.com/textcreationpartnership
}
\keyword{datasets}

