% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/richardson.R
\docType{data}
\name{richardson}
\alias{richardson}
\title{Data - Collection: Novels of Samuel Richardson}
\format{A docList with an index of 33 items.}
\usage{
richardson
}
\description{
A \code{docList} object containing metadata for a sample dataset
of Samuel Richardson's novels.
}
\section{What it contains}{

\itemize{
\item K016527.001. 1753. The history of Sir Charles Grandison.
\item K016527.002. 1753	The history of Sir Charles Grandison.
\item 	K016527.003. 1753	The history of Sir Charles Grandison.
\item 	K016527.004. 1753	The history of Sir Charles Grandison.
\item 	K016527.005. 1753	The history of Sir Charles Grandison.
\item 	K016527.006. 1753	The history of Sir Charles Grandison.
\item 	K016527.007. 1753	The history of Sir Charles Grandison.
\item 	K051638.001. 1748	Clarissa: Or, the history of a young lady.
\item 	K051638.002. 1748	Clarissa: Or, the history of a young lady.
\item 	K051638.003. 1748	Clarissa: Or, the history of a young lady.
\item 	K051638.004. 1748	Clarissa: Or, the history of a young lady.
\item 	K051638.005. 1748	Clarissa: Or, the history of a young lady.
\item 	K051638.006. 1748	Clarissa: Or, the history of a young lady.
\item 	K051638.007. 1748	Clarissa: Or, the history of a young lady.
\item 	K089369.003. 1741	Pamela: or, virtue rewarded.
\item 	K089369.004. 1741	Pamela: or, virtue rewarded.
\item 	K089370.001. 1741	Pamela: or, virtue rewarded.
}
}
\examples{
# Load the collection, then import the texts
data(richardson)
dt = importTexts(richardson)
}
\references{
https://github.com/textcreationpartnership
}
\keyword{datasets}

