% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setStopwords.R
\name{setStopwords}
\alias{setStopwords}
\title{Create or change a vector of stopwords}
\usage{
setStopwords(filepath, delimiter = ",")
}
\arguments{
\item{filepath}{A path to the file that will be converted.}

\item{delimiter}{The kind of character used to delimit your file
of stopwords. The default is "," for comma-separated files. Special
characters are included here as regular expessions, such as "\\n" for
new lines (every word is on its own line) and "\\t" for tabs.}
}
\description{
Create or change a vector of stopwords
}
\examples{
stopwords = setStopwords("~/Desktop/stopwords.txt",delimiter=",")
}

