% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup sample docLists on new machine}
\usage{
setup(object)
}
\arguments{
\item{object}{The name of the sample \code{docList} to be set up. Possible
values are "natlaw" and "richardson".}
}
\value{
A new \code{docList} object named "dl".
}
\description{
Setup sample docLists on new machine
}
\examples{
data(natlaw)
setup(natlaw)
}

