% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Data - stopwords}
\format{A character vector with 807 common words (including common typographical errors).}
\usage{
stopwords
}
\description{
A dataset containing a vector of common stopwords, including many that
are appropriate for seventeenth-century prose.
}
\section{Stopwords}{

¹,³,a,à,â,ä,ã,å,ã²,ã³,aa,aâ,aã,ab,able,about,ac,across,ad,æ,ae,æ²,af,after,ag,against,ah,
ai,aì,ak,al,all,almost,alphabet,also,am,amon,an,and,any,ao,ap,aq,ar,are,as,at,au,aw,ax,
ay,az,b,ba,bâ,bä,bã,bb,bc,bd,be,because,bee,been,before,being,bi,bl,bn,bo,both,bp,br,bs,
bt,bu,but,by,c,ca,câ,cä,cã,can,cannot,cb,cc,cd,ce,ch,ci,ck,cl,cn,co,cou,could,cp,cr,cs,ct,
cu,cy,d,da,dâ,dä,dã,db,dc,dd,de,dear,dg,di,did,dl,dn,do,doe,does,doth,dr,ds,du,duplicate,
dw,dy,e,ê,ea,eâ,eã,eb,ec,ed,ee,ef,eg,eh,ei,eì,either,ek,el,else,em,en,eo,ep,er,es,et,eu,
ev,ever,every,ew,ex,ey,ez,f,fa,fâ,fä,fã,få,fe,ff,fi,fn,fo,for,fr,from,ft,fu,fx,fy,g,ga,
gâ,gä,gã,ge,get,gg,gh,gi,gl,gn,go,got,gr,gs,h,ha,hâ,hä,hã,hå,had,has,hath,have,he,hee,her,
hers,hh,hi,him,his,hm,hn,ho,how,however,hs,ht,hu,hy,i,ì,î,ï,ia,iâ,iã,ib,ibid,ic,id,ie,if,
ig,ii,iì,iii,ij,il,im,in,into,io,ip,ir,is,it,its,iu,iv,ix,iy,j,ja,jâ,jå,je,jf,jn,jo,js,jt,
ju,k,kâ,kã,ke,kk,kn,ks,kt,ky,l,la,lâ,lä,lã,latin,lb,ld,le,least,let,lf,li,lib,like,likely,
lk,ll,lm,lo,ls,lt,lu,lw,ly,m,ma,mâ,mä,mã,må,made,make,many,may,mb,me,mi,might,mm,mn,mo,more,
most,mp,mr,mrs,ms,mt,mu,much,must,my,n,na,nâ,nä,nã,nå,nc,nd,ne,neither,ng,ni,nì,nm,nn,no,
non,nor,not,now,ns,nt,nu,nx,ny,o,oâ,oã,ob,oc,od,oe,of,off,often,og,oh,oì,ok,ol,om,on,one,
onely,only,oo,op,or,os,ot,other,ou,our,out,ov,over,ow,own,ox,oy,p,pa,pâ,pä,pã,på,page,pe,
ph,pi,pl,pn,po,pp,pr,ps,pt,pu,py,q,qa,qâ,qã,qd,qe,qq,qu,r,ra,râ,rä,rã,rå,rather,rb,rd,re,rf,
rg,ri,rm,rn,ro,rp,rr,rs,rt,ru,ry,s,sa,sâ,sä,sã,så,sæ,said,same,say,says,sb,sc,sd,se,sect,sh,
shall,she,shee,shou,should,si,sì,since,sm,so,some,sp,sr,ss,st,su,such,sw,t,ta,tâ,tä,tã,tå,
td,te,th,than,that,the,thee,their,them,then,there,these,they,thir,this,those,thou,though,
thus,thy,ti,tis,to,too,tr,ts,tt,tu,twas,ty,tz,u,ua,uâ,uã,ub,ud,ue,uf,ui,uì,uk,ul,um,un,
under,unto,uo,up,upon,ur,us,ut,uu,ux,uz,v,va,vâ,vä,vã,vå,vc,ve,very,vi,vii,viii,vl,vn,vo,vp,
vr,vs,vt,vu,vv,vy,w,wa,wâ,wã,wants,was,we,were,wh,what,when,where,which,while,who,whom,why,
wi,will,with,within,without,wm,wn,wo,wou,would,wr,ws,wt,ww,wx,x,xâ,xã,xf,xi,xii,xiii,xiv,
xix,xl,xo,xr,xt,xv,xvi,xvii,xviii,xx,y,ya,yâ,yã,yd,ye,yet,yl,ym,yn,yo,you,your,yr,ys,yt,yx,
yy,z,za,zâ,zã,ze,zh,zi,zo,zz,av,cv,cx,fb,fd,fp,gc,gd,gt,gu,hf,hr,î¹,ik,ka,kl,lc,lh,ln,lp,lr,
lv,lx,nk,np,oa,pc,qo,qs,rh,rk,rl,sn,sy,sz,tb,tc,tl,ug,uy,vê,vj,vm,xa,xc,yb,yc,yf,yi,zu,á,è,
í,ayd,wel,gat,mar,dun,tho,vve,loc,lev,lxx,hat,hoc,maâ,pow,con,pro,seq,cal,per,mis,noâ,des,
dan,bin,ere,ier,yee,mee,vse,ans,coì,wil,est,quod,therefore,another,having,here,sq,uh,ki,fk,
fg,jj,kh,xy,vh,dt,fc,kf,¹â,cf,dì,fm,gf,gm,gy,jã,mf,rq,sk,xe,_ã,ã¹,aê,bh,bq,bv,cg,df,eq,gp,
gq,hc,hd,ih,lg,mq,mv,pf,pq,qi,qm,qr,rv,tw,uc,vì,vw,wl,xd,xn,²,ð,ë,sed,quae,vel,esse,sunt,aut,
nec,quam,ver,etiam,enim,autem,quo,ejus,oft,goe,quia,sic,tamen,haec,atque,contra,inter,deut,
quid,sub,ita
}
\keyword{datasets}

