% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tcp.R
\docType{data}
\name{tcp}
\alias{tcp}
\title{Data - Index to the TCP Collection}
\format{A data frame with 61,315 rows and 10 columns.
\describe{
   \item{TCP}{id number, Text Creation Partnership}
   \item{EEBO}{id number, Early English Books Online}
   \item{VID}{(id number)}
   \item{STC}{id numbers, ESTC and Wing}
   \item{Status}{factor, levels = "Free", "Restricted"}
   \item{Author}{Author}
   \item{Date}{Date}
   \item{Title}{Title}
   \item{Terms}{Subject headings}
   \item{Pages}{Page count}
}}
\usage{
tcp
}
\description{
A dataset containing the Text Creation Partnership index.
}
\references{
https://github.com/textcreationpartnership/Texts
}
\keyword{datasets}

