% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tcpDownload.R
\name{tcpDownload}
\alias{tcpDownload}
\title{Download TCP files from Github}
\usage{
tcpDownload(results)
}
\arguments{
\item{results}{A dataframe, usually the search results generated using
              \code{searchTCP}. Must contain a column of TCP numbers with the
              name 'TCP'.}
}
\description{
Download raw XML files from the Text Creation Partnership's Github repository.
}
\section{To download the entire TCP corpus}{

If you'd like to download the entire TCP corpus, run two commands:
\enumerate{
    \item \code{data(tcp)} (Activates the TCP data in your R environment.)
    \item \code{tcpDownload(tcp)} (Begins the download.)
}

Be careful: the entire TCP is several GBs, so the download will take awhile.
}
\examples{
tcpDownload(results)
}

