% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tei2r.R
\docType{package}
\name{tei2r}
\alias{tei2r}
\alias{tei2r-package}
\title{tei2r: Import and Parse TEI documents in R}
\description{
The \code{tei2r} package imports collections of plain-text or TEI-encoded
files into the R environment and makes them available for quantitative
analysis.
}
\section{What it does}{

The basic purpose of the \code{tei2r} package is to help users get their
TEI-encoded files into R. It is specifically designed for use with files
generated by the Text Creation Partnership, providing easy access to
thousands of early modern and eighteenth-century texts.

For a tutorial that walks through the basics, check out the introductory
vignette, "Introduction: Getting Started with tei2r."

\strong{Basic operations:}
\itemize{
  \item \code{\link{buildDocList}}: Represents your collection of documents
  \item \code{\link{importTexts}}: Moves texts into R
  \item \code{\link{buildConcordance}}: Parses the documents for keywords-in-context
}

\strong{Getting Started:}
\itemize{
  \item \code{\link{tcpSearch}}: Search the TCP index
  \item \code{\link{tcpDownload}}: Download files from the TCP's Git repository
}

\strong{Sample Data:}
\itemize{
  \item \code{\link{natlaw}}: A collection of 33 documents related to 17th-cent
                              political theory.

  \item \code{\link{richardson}}: The novels of Samuel Richardson.
}
}
\author{
Authors:
Michael Gavin and Travis Mullen
}

