% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tfidf_test.R
\name{tfidf}
\alias{tfidf}
\alias{tfidfAssociation}
\alias{tfidfConcordance}
\title{This function calculates the Term Frequency Inverse Document
Frequency in a variety of situations, including: a single term
across all documents, multiple terms, all terms, all terms (with limit)
for \code{docFrequencies}, \code{docConcordance}, or \code{docAssociations}
objects.}
\usage{
tfidf(df, term = "", all = F, limit = 1000)

tfidfAssociation(da, term = "", all = F)

tfidfConcordance(dc, term, all)
}
\arguments{
\item{df}{A \code{tei2r} object of class \code{docFrequencies},
               \code{docConcordance}, or \code{docAssociations}. This
               object contains the terms to calculate the \code{tfidf}
               for.}

\item{term}{The term or list of terms (\code{c("termOne", "termTwo")})
               to calculate \code{tfidf} for.}

\item{all}{A boolean value that tells the function whether to calculate
               \code{tfidf} for all terms or not (limited to 1000 by default).}

\item{limit}{A numerical value that tells the function how many terms to limit
               \code{all} to.  Defaults to 1000.}
}
\value{
tf.idf A list of numerical values for each \code{term} that represents
               their \code{tfidf} value in each document.
}
\description{
This function calculates the Term Frequency Inverse Document
Frequency in a variety of situations, including: a single term
across all documents, multiple terms, all terms, all terms (with limit)
for \code{docFrequencies}, \code{docConcordance}, or \code{docAssociations}
objects.
}
\section{What It Does}{


The Term Frequency Inverse Document Frequency is determined
by multiplying the frequency of a term in a document by the
log of the total number of documents divided by the number
of documents in which the term appears:

f(term) * log(numDocuments/numDocumentsWithTerm)
}
\examples{
tf = tfidf(df, term="just")
tf = tfidf(df, term=c('just', 'right'))
tf = tfidf(df, all=T)
tf = tfidf(df, all=T, limit=5000)

tf = tfidf(dc, all=T)

tf = tfidf(da, all=T)
}

