% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/useWizard.R
\name{useWizard}
\alias{useWizard}
\title{This function walks the user through the process of building a
\code{docList} object.  It preforms the same functionality of
the \code{buildDocList} function, but is interactive and works
in a more simple way.  This function should \emph{only} be called
by using the \code{wizard = T} parameter of the \code{buildDocList}
function.}
\usage{
useWizard(dl, directory = "")
}
\arguments{
\item{dl}{The empty \code{docList} object that is created in
the \code{buildDocList} function and that is constructed
in this function.  The object is returned to the
\code{buildDocList} function.}

\item{directory}{The directory in which the corpus files are stored.  Passing
a value in for this parameter is optional and saves the user
from being asked for their directory if they have already
provided it.}
}
\description{
This function walks the user through the process of building a
\code{docList} object.  It preforms the same functionality of
the \code{buildDocList} function, but is interactive and works
in a more simple way.  This function should \emph{only} be called
by using the \code{wizard = T} parameter of the \code{buildDocList}
function.
}

