% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xmlToText.R
\name{xmlToText}
\alias{xmlToText}
\title{This function converts the xml file found at \code{filepath} to
plain text and, optionally, stores it at \code{target}.}
\usage{
xmlToText(filepath, target = NULL)
}
\arguments{
\item{filepath}{A string that holds the path to the xml file
                 to convert to text.}

\item{target}{A string that holds the path at which to store
                 the converted text file.}
}
\value{
text     The plain text extracted from the xml file.
}
\description{
This function converts the xml file found at \code{filepath} to
plain text and, optionally, stores it at \code{target}.
}
\examples{
fp = "path/to/file/text.xml"
text = xmlToText(fp)
text = xmlToText(fp, target="/alternate/path/for/new/file/")
}

