% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textCleanup.R
\name{cleanup}
\alias{cleanup}
\title{Convert a plain-text or TEI document into a character vector}
\usage{
cleanup(filepath, removeCaps = TRUE, stopwords = dl@stopwords,
  removeStopwords = TRUE)
}
\arguments{
\item{filepath}{A path to the file that will be converted.}

\item{removeCaps}{A logical condition. If "TRUE", all words will be converted to lower case.}

\item{removeStopwords}{A logical condition. If "TRUE", all words contained in a pre-defined vector of stopwords will be excluded.}

\item{normalizeLongS}{A logical condition. If "TRUE", all instances of '∫' and 'ſ' will be converted to 's'.}
}
\description{
Convert a plain-text or TEI document into a character vector
}
\examples{
locke.path = "~/Desktop/locke2ndTreatise.txt"
cleanup(locke.path, removeCaps = TRUE, removeStopwords = FALSE)
}

