% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/textCleanup.R
\name{cleanup}
\alias{cleanup}
\alias{teiTextCleanup}
\alias{textCleanup}
\title{Convert a plain-text document into a character vector}
\usage{
textCleanup(filepath, removeCaps = TRUE, stopwords = dl@stopwords,
  removeStopwords = TRUE)

teiTextCleanup(filepath, stopwords)
}
\arguments{
\item{filepath}{A path to the file that will be converted.}

\item{removeCaps}{A logical condition. If "TRUE", all words will be converted to lower case.}

\item{removeStopwords}{A logical condition. If "TRUE", all words contained in a pre-defined vector of stopwords will be excluded.}
}
\description{
Convert a plain-text document into a character vector
}
\examples{
locke.path = "~/Desktop/locke2ndTreatise.txt"
textCleanup(locke.path, removeCaps = TRUE, removeStopwords = FALSE)
}

