% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docConcordance.R
\docType{class}
\name{docConcordance-class}
\alias{docConcordance}
\alias{docConcordance-class}
\title{An S4 class that holds a concordance for one term across document collections}
\description{
\code{docConcordance} creates a special \code{tei2r} object called
}
\section{Slots}{

\describe{
\item{\code{directory}}{A string ...}

\item{\code{indexFile}}{A string [NOTE: THIS SHOULD BE CALLED indexFile.}

\item{\code{term}}{A string ...}

\item{\code{context}}{An integer [This probably doesn't matter, but it should
be an integer rather than a numeric]}

\item{\code{concordance}}{A list...}
}}
\section{What it does}{

Here we can paste longer descriptions of what each function does.
}

