% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docFrequencies.R
\docType{class}
\name{docFrequencies-class}
\alias{docFrequencies}
\alias{docFrequencies-class}
\title{An S4 class that contains and manages word-frequency data}
\description{
\code{docFrequencies} creates a special \code{tei2r} object that holds word-frequency data for each document.
}
\section{Slots}{

\describe{
\item{\code{directory}}{A string that gives the filepath to the main directory}

\item{\code{indexFile}}{A string providing the filepath to the index (usually a .csv file)}

\item{\code{raw}}{A list of named vectors, showing the raw frequency of each word in each document}

\item{\code{proportional}}{A list of vectors showing the proportional frequency of each word}

\item{\code{vocabulary}}{A vector that holds the frequencies of each word across
the whole collection.}

\item{\code{proportionalVocab}}{A vector that holds the proportional frequency of
                        of each word over the entire collection.}
}}
\section{What it does}{

The \code{docFrequencies} object is a structured list that hold word frequencies
for each document in the collection, as well as summary data about the collection
as a whole.
}

