% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docList.R
\docType{class}
\name{docList-class}
\alias{docList}
\alias{docList-class}
\title{An S4 class to represent a document collection.}
\description{
\code{docList} returns a special \code{tei2r} object that contains
all of the necessary information to begin working with a corpus of
text in R.
}
\section{Slots}{

\describe{
\item{\code{directory}}{A string that gives the filepath to the main directory (folder)}

\item{\code{filenames}}{A vector containing all of the filenames for the documents in
the corpus.}

\item{\code{paths}}{A vector containing the full path to all files in the corpus.}

\item{\code{indexFile}}{A string that gives the filepath to the index file for the
corpus.  This file should house the meta-data for each file
in the corpus.}

\item{\code{index}}{A data frame that holds the meta-data for each document in
the corpus.  This data frame is created by reading the file
found at \code{indexFile.}}

\item{\code{stopwordsFile}}{A string that gives the filepath to the file that contains
a comma seperated list of words to be removed during text
cleanup.}

\item{\code{stopwords}}{A vector derived from the \code{stopwordsFile} that is passed
                     to the text cleanup functions in order for them to be removed
                     from the text.}
}}
\section{What it does}{

The \code{docList} is the foundaiton of the \code{tei2r}
package and, as such, should be the first object created when working
with the package.  The object is constructed by calling the
\code{buildDocList} function.  This function builds the object by
storing the path to the corpus' files (\code{directory}), the
file containing the corpus' meta-data (\code{indexFile}), and
the stopwords file (\code{stopwordsFile}).  From these pieces
of information, the function automatically determines the
\code{filenames} and \code{paths} for the corpus' files.
}

