% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docTexts.R
\docType{class}
\name{docTexts-class}
\alias{docTexts}
\alias{docTexts-class}
\title{An S4 class that contains and manages text data from document collections}
\description{
\code{docTexts} creates a special \code{tei2r} object called
}
\section{Slots}{

\describe{
\item{\code{directory}}{A string ...}

\item{\code{index}}{A string [NOTE: THIS SHOULD BE CALLED indexFile.}

\item{\code{text}}{A list ...}
}}
\section{What it does}{

Here we can paste longer descriptions of what each function does.
}

