% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findFilenames.R
\name{findFilenames}
\alias{findFilenames}
\title{Placeholder description}
\usage{
findFilenames(dl, directory)
}
\arguments{
\item{dl}{}

\item{directory}{}
}
\description{
Placeholder description
}
\examples{
Placeholder
}

