% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aliases.R, R/getDocAssociations.R
\name{associations}
\alias{associations}
\alias{getDocAssociations}
\title{Determine the associations (frequency of co-occurrance) for the
\code{keyword} that was chosen when making the \code{docConcordance}.}
\usage{
associations(dc, df)

getDocAssociations(dc, df)
}
\arguments{
\item{dc}{A \code{docConcordance} object}

\item{df}{A \code{docFrequencies} object}
}
\description{
Determine the associations (frequency of co-occurrance) for the
\code{keyword} that was chosen when making the \code{docConcordance}.
}
\section{What it does}{

Gets the associations across the collection for a given term.  This
function uses a \code{docConcordance} object to determine the frequency
with which words appear with the term you selected when calculating
the concordance.  It returns the raw frequencies for the term across
each document in the collection, and it also provides the total
across the collection as a whole.
}

