% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aliases.R, R/getDocAssociations.R
\name{associations}
\alias{associations}
\alias{getDocAssociations}
\title{Determine the associations (frequency of co-occurrance) for the
\code{term} that was chosen when making the \code{docConcordance}
in both raw count and proportion across the corpus.}
\usage{
associations(dc, df)

getDocAssociations(dc, df)
}
\arguments{
\item{dc}{}

\item{df}{}
}
\description{
Determine the associations (frequency of co-occurrance) for the
\code{term} that was chosen when making the \code{docConcordance}
in both raw count and proportion across the corpus.
}

