% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aliases.R, R/getDocConcordance.R
\name{getConcordance}
\alias{getConcordance}
\alias{getDocConcordance}
\title{Get the concordance for \code{term} within \code{context} across
all texts in \code{dt}.}
\usage{
getConcordance(dt, term, context)

getDocConcordance(dt, term, context)
}
\arguments{
\item{dt}{A \code{docTexts} object that contains your collection
of texts.  These are the texts that will be used when
searching for \code{term}s concordance (the words that
appear within a certain \code{context} of \code{term}.}

\item{term}{The term that you are interested in finding the
concurring words for across the text.}

\item{context}{The number of words on each side of the
\code{term}.}
}
\description{
Get the concordance for \code{term} within \code{context} across
all texts in \code{dt}.
}
\examples{
dc = getDocConcordance(dt, "just", 5)
}

