% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aliases.R, R/getDocFrequencies.R
\name{getFrequencies}
\alias{getDocFrequencies}
\alias{getFrequencies}
\title{Get the frequency of all words in each text in \code{dt} and
optionally of all words in the corpus.}
\usage{
getFrequencies(dt)

getDocFrequencies(dt, limit = 0)
}
\arguments{
\item{dt}{A \code{docTexts} object that contains the full
text of each document in your corpus.}

\item{limit}{A numeric value that limits the number of words
               to find the frequency of in each document in
               \code{dt}. Defaults to all terms.}
}
\value{
df     a \code{docFrequencies} object with frequency
               data for the corpus.
}
\description{
Get the frequency of all words in each text in \code{dt} and
optionally of all words in the corpus.
}
\examples{
df = getDocFrequencies(dt)
df = getDocFrequencies(dt, 500)
df = getDocFrequencies(dt=dt, limit=1000)
}

