% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aliases.R, R/getDocTexts.R
\name{getTexts}
\alias{getDocTexts}
\alias{getTexts}
\title{This function collects, cleans up (removes stopwords), and stores
the text of the corpus' documents in a single object in which the
text is labeled by the corresponding \code{id} in the \code{index}
file.  This is usually the filename sans extension or the \code{tcp}
number.}
\usage{
getTexts(dl)

getDocTexts(dl)
}
\arguments{
\item{dl}{The \code{docList} object that contains the \code{index} with
          the \code{paths} to the files that house the text.}
}
\value{
dt The \code{docTexts} object that contains the texts of the corpus,
           the path to the \code{indexFile} and the original \code{directory}
           that the \code{docList} object was constructed with.  This information
           enables the user to keep track of which objects are associated with
           which corpus.
}
\description{
This function collects, cleans up (removes stopwords), and stores
the text of the corpus' documents in a single object in which the
text is labeled by the corresponding \code{id} in the \code{index}
file.  This is usually the filename sans extension or the \code{tcp}
number.
}
\examples{
dt = getDocTexts(dl)
}

