% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Data - stopwords}
\format{A character vector with 300 common words.}
\usage{
stopwords
}
\description{
Stopwords:
have,has,d,his,which,l,ibid,s,sect,those,upon,one,a,able,about,across,after,
all,almost,also,am,amon,an,and,any,are,as,at,be,because,been,but,by,can,
cannot,could,dear,did,do,does,either,else,ever,every,for,from,get,got,had,
he,her,hers,him,his,how,however,i,if,in,into,is,it,its,least,let,like,likely,
may,me,might,most,must,my,neither,no,nor,not,of,off,often,on,only,or,other,
our,own,rather,said,say,says,she,should,since,so,some,than,that,the,their,
them,then,there,these,they,this,tis,to,too,twas,us,wants,was,we,were,what,
when,where,which,while,who,whom,why,will,with,would,yet,you,your,thir,n,ye,t,
mr,p,ii,iii,ed,h,th,such,o,1,2,3,4,5,6,7,8,9,10,11,12,er,hath,more,qu,very,
much,m,r,sh,st,ne,c,ad,non,lib,de,unto,shall,have,has,d,his,which,l,ibid,s,
sect,those,upon,one,a,able,about,across,after,all,almost,also,am,amon,an,and,
any,are,as,at,be,because,been,but,by,can,cannot,could,dear,did,do,does,either,
else,ever,every,for,from,get,got,had,he,her,hers,him,his,how,however,i,if,in,
into,is,it,its,least,let,like,likely,may,me,might,most,must,my,neither,no,nor,
not,of,off,often,on,only,or,other,our,own,rather,said,say,says,she,should,
since,so,some,than,that,the,their,them,then,there,these,they,this,tis,to,too,
twas,us,wants,was,we,were,what,when,where,which,while,who,whom,why,will,with,
would,yet,you,your,thus
}
\details{
A dataset containing a vector of common stopwords, including many that
are appropriate for seventeenth-century prose.
}
\keyword{datasets}

