% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tei2r.R
\docType{package}
\name{tei2r}
\alias{tei2r}
\alias{tei2r-package}
\title{tei2r: Import and Parse TEI documents in R}
\description{
The \code{tei2r} package imports collections of plain-text or TEI-encoded
files into the R environment and makes them available for quantitative
analysis.
}
\section{What it does}{

The basic purpose of the \code{tei2r} package is to help users get their
TEI-encoded files into R.
}
\author{
Authors:
Michael Gavin and Travis Mullen
}

